Extended information reported by get-ballance.
The core module extended with procedures for checking if a symbol is a
currency or a stock, and with a procedure for calculating present NAV
(net asset value) of symbol(s) in the DB.

The get-ballance script was extended to also report NAV for DB stocks
and a total DB NAV (incl. convertion to a grand total in a base
currency).

Fixed the BASH script for setting up a clean live environment.