Added IRR for performance and updated to Finance::Quote 1.47.
The xfrs module and get-stock-irr script were updated to the latest
Finance::Quote (1.47) and use AlphaVantage source to query stocks and
currencies. The proper use expects the ALPHAVANTAGE_API_KEY environment
variable to be set.

The get-performance script was updated to also report IRR for individual
stocks and cumulative for all stocks denominated in the same currency.