Implemented `get-quotes` in Python.
This Python implementation is different from the Perl one. The Perl
version is dated and was coded for debugging in the early days of development, then
no longer used. Hence the Perl version did not really reported just
the quotes.

The Python version now implements the more or less expected function of
the `get-quotes` script.