Progressed on `import-xfrs.py`.
This commit adds parsing the input file records into the structured form
to be entered into DB.

The commit started standardization of FX (currency pair) representation.
In the past, the FX quotes were not in the DB. The new approach pivots to
using the Yahoo Finance tickers (e.g. `USDEUR=X`) in DB `quotes` table.
This makes it 1:1 correspondence between the `yfinance` quote and should
guarantee unique recognition and recovery of the FX pair.