Improved getting cached quotes.
With delisted stocks, `yfinance` quotes would fail and some of the
scripts might yield weird results. Improved queries to cached/DB
quotes to return the quote on date of delisting, if captured in DB.

Also extended `cache-quotes.py` to query and cache historical prices.
This would not work for delisted stocks, though, and these would need
to be imported into DB manually.